Correção nos métodos delete em controllers e melhoria de código

+++ src/controllers/PedidoController.js
@@ -5,9 +5,9 @@ class PedidoController {
 	static async pegaTodosPedidos(req, res) {
 		try {
 			const todosPedidos = await database.Pedidos.findAll();
-			return res.status(200).json(todosPedidos);
+			return res.status(200).send(todosPedidos);
 		} catch (error) {
-			return res.status(400).json(error.message);
+			res.status(400).send(error.message);
 		}
 	}
 
@@ -16,9 +16,9 @@ class PedidoController {
 		const { id } = req.params;
 		try {
 			const umPedido = await database.Pedidos.findOne({ where: { id: Number(id) } });
-			return res.status(200).json(umPedido);
+			return res.status(200).send(umPedido);
 		} catch (error) {
-			return res.status(400).json(error.message);
+			res.status(400).send(error.message);
 		}
 	}
 
@@ -27,9 +27,9 @@ class PedidoController {
 		const novoPedido = req.body;
 		try {
 			const novoPedidoCriado = await database.Pedidos.create(novoPedido);
-			return res.status(201).json(novoPedidoCriado);
+			return res.status(201).send(novoPedidoCriado);
 		} catch (error) {
-			return res.status(400).json(error.message);
+			res.status(400).send(error.message);
 		}
 	}
 
@@ -39,9 +39,9 @@ class PedidoController {
 		const pedidoAtualizado = req.body;
 		try {
 			await database.Pedidos.update(pedidoAtualizado, { where: { id: Number(id) } });
-			return res.status(200).json(pedidoAtualizado);
+			return res.status(200).send(pedidoAtualizado);
 		} catch (error) {
-			return res.status(400).json(`Não foi possivel atualizar o pedido: ${error.message}`);
+			res.status(400).send(`Não foi possivel atualizar o pedido: ${error.message}`);
 		}
 	}
 
@@ -49,10 +49,15 @@ class PedidoController {
 	static async deletaPedido(req, res) {
 		const { id } = req.params;
 		try {
+			const umPedido = await database.Pedidos.findOne({ where: { id: +id } });
+			if (!umPedido) return res.status(404).send(`Pedido de id ${id} não existe`);
+			const cliente = await database.Clientes.findOne({ where: { id: +id } });
 			await database.Pedidos.destroy({ where: { id: Number(id) } });
-			return res.status(200).json(`mensagem: O pedido com o id ${id} foi deletado com suceso`);
+			return res
+				.status(200)
+				.send(`mensagem: O pedido id ${id} de ${cliente.nome} foi deletado com suceso`);
 		} catch (error) {
-			return res.status(400).json(error.message);
+			res.status(400).send(error.message);
 		}
 	}
 }
